requirements.txt
Django==4.2.15
djangorestframework_simplejwt==5.3.0
drf_yasg==1.21.7
PyJWT==2.8.0
python-decouple==3.8

gunicorn

